requirements.txt
ultralytics
opencv-python
numpy
fast-plate-ocr[onnx] # Cross-platform
mysql-connector-python
gpiozero
pigpio

# Python 3.13.5 64-bit